Update SceneServiceInfo.js
Hi Thorsten,
please include information regarding lodtype per the spec update sent out in feb for the Layer object of a 3DSceneServiceInfo. This document, a.ka. 'SceneServer' rest resource is the first document the client retrieves and will be beneficial to the know the lodType to decide access patterns etc.. Thanks